requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
imutils==0.5.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.1
olefile==0.46
opencv-python==3.4.3.18
Pillow==4.1.0
requests==2.13.0
visitor==0.1.3
Werkzeug==0.14.1
